Fix Vue Auth navigation link
When trying to customize Sign Up with custom fields and attributes, navigation links do not point to the correct anchor. That can mislead people from going straight into signUpFields where the wrong configuration will be silently accepted. 

This PR is a minor fix to the navigation so one doesn't miss signUpField which is the attribute they're looking for in the first place.