Updating auto-refresh language
From what I can see in the source code, the language here is confusing.  When is "appropriate"?  Does that mean just before it expires?  When the user calls Auth.currentSession()? 

I think it makes it clearer that the refresh will take place only when a user consuming the API makes calls into the API itself.