Fix graphql example about Owner Authorization
To be honest, I'm not sure if I ran into a bug or if the example is wrong. But if I follow the example (Ownership with create mutations), I get an error message that I am not sending values for the "editors" field ("Variable 'input' has coerced Null value for NonNull type '[String]!"). The ownership fields should be filled automatically unless told explicitly not to do so. The example worked for me when I marked the "editors" field as no longer required.