Add instruction for existing CRA projects
It isn't immediately clear that you can simply add Amplify to an existing CRA project. The instructions suggest you must make a new project. I've added an option 1 / option 2 to the beginning of the installation docs to make it clear that if you already have an existing project you just have to cd into it and run amplify init.