Update Android Getting Started intro/dependencies (#1429)
The Android Getting Started guide made reference to the beta version of the code-generator plugin, and it made reference to an out of date version of the Android Gradle Plugin. Both versions have been updated to the latest.

The wording in the first paragraph of the Getting Started guide has also received another pass, towards improving its clarity.