Optimize Android Datastore Examples for Screen Width
On a MacBook, there is very little width for the main content of the
documentation, on the V2 docs site.

To present a concise and readable example to the user, a balance needs
to be found between descriptive and meaningful phrasing, and getting to
the point.

Requiring the user to scroll left/right to understand basic examples
isn't helping.

Some basic variable renaming / formatting changes are made to push the
content of the examples to the left of the screen.