Add SDK Setup Options
In our previous docs site, there was a section of the SDK docs for
developers who wish to use the low level clients. This change moves
that over and adds it to the SDK docs under a "Configuration" header.

Note: I did not proofread these documents or check them for currency,
rather this is a straight port of what is in master.