Updated descriptions of Admin Queries API routes.
All the current descriptions of the Admin Queries API routes are "Adds a user to a specific Group". This description does not match what the each of the API routes does.

I updated the description for the each of the routes,  describing as accurately as possible what each of the routes does.