Removing choice in getting started
I think it was confusing to read the getting started guide and have it say "choose this" or "this".  I think it makes sense just to see a full example soup to nuts.  I consider getting started as a place we need to hold the customers hand and we should take them through a 'hello world' example.