Reload page on changes in development
For whatever reason, the default reload behavior of Hot Module
Replacement is not working within this project which results in the
content author being forced to tediously reload on changes. This change
configures the behavior to `pageReload` which will do a full refresh of
the page on rebuild.

See: https://stenciljs.com/docs/dev-server#dev-server-config