Hide copy button for console snippets
Right now, we show the copy button for output that's meant to be
expository rather than a block we expect a developer to copy and paste.
The most common example is CLI output: while this is useful to see to
understand what options should be selected, there's no value in copying
it into your clipboard.

This change suppresses the copy button if the language is set to
"console" which is defined in Linguist. See: https://bit.ly/2XsAkBq