Suppress copy button when unnecessary (#1608)
We have many places on the docs site where a user can copy a
pre-formatted code block that's actually console output. While the copy
button suggests to the user they should copy it, in fact this action is
valueless. In https://github.com/aws-amplify/docs/pull/1557, a change
was made to hide the copy button if the language `console` was used.
This change goes through the docs and makes the following changes:

- Moves from bash, terminal, sh, and others to console if the user is
  not intended to copy and paste the pre-formatted block
- Ensure that items marked as `bash` do not have a leading prompt to
  allow the user to actually copy and paste it

Resolves #1596.

Co-authored-by: Rene Brandel <brene@users.noreply.github.com>