Fix copy button to work more than once
Users currently can't hit the copy button more than once. On the second
press it will say "failed to copy."

It appears we're using a ref to keep a handle on the div to grab its
textContent. When the component unmounts, React calls the binding with
null to prevent memory leaks.  When it mounts, React will call the
binding again with the element.  Apparently in some browsers these
callbacks happen out of order, such that we override our reference with
null. The React documentation lightly hints at this possibility:
https://reactjs.org/docs/refs-and-the-dom.html#callback-refs. This
change ignores calls with null as an argument