Fix Angular hosting DistributionDir (#1744)
* Update amplify init instructions & fix app name

Set the correct Distribution directory path, and also change the name of the project to myAmplifyProject, which is the name used with ng new

* Add instructions for AccessDenied error

The Angular tutorial project uses this page, as do other projects.  Due to a change in the distribution directory used with recent Angular versions, hosting produces an XML document with the error AccessDenied, and no clue to help the frustrated student.  Therefore, add a clue that will lead to the problem being solved.  Angular is not referenced here because this page is used for several frameworks, not just Angular.

Co-authored-by: Rene Brandel <brene@users.noreply.github.com>
Co-authored-by: Sam Martinez <samlmar@amazon.com>