Improve tutorial wording for react todos
Before the changes in this commit, the instructions about updating `src/App.js` were ambiguous. It looks like the instructions want you to do a full replacement of the file with the new content they show, but the new content lacks the previous lines to import Amplify and awsExports and configure Amplify appropriately.

This commit adds those lines to the `src/App.js` and makes it explicit that a full replacement is what is instructed.