Don't put Existing Resources section on JS
The self-closing tag used for inline-fragment is invalid which causes
this page to appear in the menu for JavaScript where we have no content.
This change puts in an explicit close tag which continues to render the
content and ensures it doesn't appear in the menu for JavaScript where
there isn't any content.