Redirect Flutter users to platform docs (#2277)
Instead of showing a warning on Guides, just redirect the user to the
existing documentation if they happen to use the Flutter dropdown on
this page. This is what we do on the SDK pages for JavaScript so it fits
an existing pattern, and it eliminates unnecessary pages. This also
fixes a bug on the SDK pages where using the Flutter dropdown would
result in a 404.