fix: advertise rxjava as solution to nested android calls
In the portion of Android DataStore documentation discussing
many-to-many relationships, there is a code snippet showing a
triple-nested saved. This code is accompanied by a note "we're working
on ways to improve this experience." It is an out-of-date note. Since
the callout was written, we have improved the experience by adding
support for RxJava. The note is updated to highlight RxJava as a
possible soultion to the nested callbacks.