extract cli install code block to fragment
**Notes:**

Currently the CLI install code snippet is replicated in 15+ files. A
single change requires too many changes and it's hard to guarantee all
files are updated, correct and new files will follow the correct content
and format.

I extracted the install code block into a fragment and updated all the
occurances to the CLI install code.

**ARM / Apple M1**:

The Packaged CLI does not work on the new ARM-based architecture, the
Apple M1. I added a call-out until that issue is fixed.