Clarified Amplify CLI configuration w/o AWS steps
The instructions said to navigate to open a new terminal window and
run `amplify pull --appId ...` without mentioning that one should
be in the directory that they would like amplify cli to install their
project in. If the user does not infer this it will install in their
home directory which is very likely not their choice project directory.

I updated the instructions to tell the user to change to their desired
directory before running the `amplify pull --appId ...` command.