Android docs: uniform label for coroutine APIs (#3020)
In order for the language switcher's state management to work properly,
we need to use the same title for the Flow/Coroutine group.

Previously, some snippets said "Kotlin - Coroutines (Beta)", while
others said "Kotlin - Flow (Beta)". While that may be an optimal
distinction from the user's standpoint, it's better to have the
persistent toggle working in the switcher.