Add note for using structured types with functions (#3128)
* Add note for using structured types with functions

As a first-time Lambda, I was unfamiliar with the expected input format for these functions. This note adds a link to the relevant Lambda documentation. While the examples in this document are helpful for the general expected design, it is difficult to understand the important parts required to implement a correct lambda resolver for a GraphQL function. I added the note to help other developers using more strongly typed languages such as Golang, Java, and C# should they encounter this error.

Thank you for considering this change!

* Re-trigger stalled CI/CD

Co-authored-by: Michael Edelman <metalman@amazon.com>