Found another typo (#3269)
Should be "with optional parameters `limit` and `where` to specify the page size and condition" instead of "limit` and `where` to specific the page size and condition"

You can get the list of items using `.paginatedList` with optional parameters `limit` and `where` to specific the page size and condition. By default, the page size is 1000.