chore(updating hosting.mdx): add explicit redirect address
In the documentation for Hosting, it states that there should be a redirect address but it does not specify what that address should be. I have created a react app with navigation controlled by react-router. I think it would be best for our customers to have that route stated explicitly without having to look through various stack overflow pages.