chore(updating hosting.mdx): add explicit redirect address (#3772)
* chore(updating hosting.mdx): add explicit redirect address

In the documentation for Hosting, it states that there should be a redirect address but it does not specify what that address should be. I have created a react app with navigation controlled by react-router. I think it would be best for our customers to have that route stated explicitly without having to look through various stack overflow pages.

* Update hosting.mdx

* Update src/pages/cli/hosting/hosting.mdx

Co-authored-by: siegerts <stephen.siegert@gmail.com>

Co-authored-by: siegerts <stephen.siegert@gmail.com>