Added migration case regarding OpenSearch 1.1
Users who update to OS after migration and choose to upgrade to 1.1 will encounter an issue with the existing searchable queries due to an indexPath issue. The indexPath should be manually updated because modifying the CLI transformers might break a few things. 

The workaround, for now, is copying the resolver to the resolver folder outside build and updating the path manually.