Update Custom Business Logic - Add model attribute example
This improves the documentation around creating a custom resolver. This shows how an attribute of a model can be used to reference another table to return a value. This also improves the documentation for the custom resolver demonstration code to show that typeName can be more than just the 3 originally demonstrated. Also include comment to explain that 'Table' is required.