docs(server-side-rendering): replace deprecated 'key' with 'path' in static rendering example
### Description of changes
Replaced the deprecated `key` parameter with `path` in the Static rendering example for React server components.

The `getUrl()` function shows the following deprecation warning:
"The key and accessLevel parameters are deprecated and may be removed in the next major version. Please use path instead."

### Related GitHub issue #, if available:
N/A

### Product(s) affected
- [x] amplify-libraries

### Platform(s) affected
- [x] JS

### Checks
- [x] Does this PR conform to the styleguide?
- [x] Are all links in MDX files using the MDX link syntax?

### Ready to merge
- [x] Ready to merge

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.